+++ crc/api/file.py
@@ -8,14 +8,19 @@ from flask import send_file
 from crc import session
 from crc.api.common import ApiError
 from crc.api.user import verify_token
-from crc.models.file import FileSchema, FileModel, File, FileModelSchema, FileDataModel, FileType
+from crc.models.file import FileSchema, FileModel, File, FileModelSchema, FileDataModel, FileType, CONTENT_TYPES
 from crc.models.workflow import WorkflowSpecModel
 from crc.services.document_service import DocumentService
 from crc.services.file_service import FileService
+from crc.services.spec_file_service import SpecFileService
 
 def to_file_api(file_model):
     """Converts a FileModel object to something we can return via the api"""
-    return File.from_models(file_model, FileService.get_file_data(file_model.id),
+    if file_model.workflow_spec_id is not None:
+        file_data_model = SpecFileService().get_spec_file_data(file_model.id)
+    else:
+        file_data_model = FileService.get_file_data(file_model.id).data
+    return File.from_models(file_model, file_data_model,
                             DocumentService.get_dictionary())
 
 
@@ -64,8 +69,8 @@ def add_file(workflow_spec_id=None, workflow_id=None, task_spec_name=None, form_
         else:
             primary = False
         workflow_spec = session.query(WorkflowSpecModel).filter_by(id=workflow_spec_id).first()
-        file_model = FileService.add_workflow_spec_file(workflow_spec, file.filename, file.content_type,
-                                                        file.stream.read(), primary=primary)
+        file_model = SpecFileService.add_workflow_spec_file(workflow_spec, file.filename, file.content_type,
+                                                            file.stream.read(), primary=primary)
     else:
         raise ApiError("invalid_file", "You must supply either a workflow spec id or a workflow_id and form_field_key.")
 
@@ -73,11 +78,13 @@ def add_file(workflow_spec_id=None, workflow_id=None, task_spec_name=None, form_
 
 
 def get_reference_file(name):
-    file_data = FileService.get_reference_file_data(name)
+    file_extension = FileService.get_extension(name)
+    content_type = CONTENT_TYPES[file_extension]
+    file_data = SpecFileService.get_reference_file_data(name)
     return send_file(
-        io.BytesIO(file_data.data),
-        attachment_filename=file_data.file_model.name,
-        mimetype=file_data.file_model.content_type,
+        io.BytesIO(file_data),
+        attachment_filename=name,
+        mimetype=content_type,
         cache_timeout=-1  # Don't cache these files on the browser.
     )
 
@@ -99,7 +106,7 @@ def set_reference_file(name):
 
     file_models = FileService.get_files(name=name, is_reference=True)
     if len(file_models) == 0:
-        file_model = FileService.add_reference_file(name, file.content_type, file.stream.read())
+        file_model = SpecFileService.add_reference_file(name, file.content_type, file.stream.read())
     else:
         file_model = file_models[0]
         FileService.update_file(file_models[0], file.stream.read(), file.content_type)
@@ -109,7 +116,7 @@ def set_reference_file(name):
 
 def add_reference_file():
     file = connexion.request.files['file']
-    file_model = FileService.add_reference_file(name=file.filename, content_type=file.content_type,
+    file_model = SpecFileService.add_reference_file(name=file.filename, content_type=file.content_type,
                                                 binary_data=file.stream.read())
     return FileSchema().dump(to_file_api(file_model))
 
@@ -126,16 +133,20 @@ def get_file_data_by_hash(md5_hash):
     filedatamodel = session.query(FileDataModel).filter(FileDataModel.md5_hash == md5_hash).first()
     return get_file_data(filedatamodel.file_model_id,version=filedatamodel.version)
 
+
 def get_file_data(file_id, version=None):
-    file_data = FileService.get_file_data(file_id, version)
+    file_model = session.query(FileModel).filter(FileModel.id==file_id).first()
+    if file_model.workflow_spec_id is not None:
+        file_data = SpecFileService().get_spec_file_data(file_id)
+    else:
+        file_data = FileService.get_file_data(file_id, version).data
     if file_data is None:
         raise ApiError('no_such_file', f'The file id you provided ({file_id}) does not exist')
     return send_file(
-        io.BytesIO(file_data.data),
-        attachment_filename=file_data.file_model.name,
-        mimetype=file_data.file_model.content_type,
-        cache_timeout=-1,  # Don't cache these files on the browser.
-        last_modified=file_data.date_created
+        io.BytesIO(file_data),
+        attachment_filename=file_model.name,
+        mimetype=file_model.content_type,
+        cache_timeout=-1  # Don't cache these files on the browser.
     )
 
 
@@ -180,7 +191,7 @@ def update_file_info(file_id, body):
 def delete_file(file_id):
     workflow_spec_id = session.query(FileModel.workflow_spec_id).filter(FileModel.id==file_id).scalar()
     if workflow_spec_id is not None:
-        FileService.delete_spec_file(file_id)
+        SpecFileService.delete_spec_file(file_id)
     else:
         FileService.delete_file(file_id)
