+++ crc/models/api_models.py
@@ -212,14 +212,13 @@ class DocumentDirectory(object):
 
 class WorkflowApi(object):
     def __init__(self, id, status, next_task, navigation,
-                 is_latest_spec, workflow_spec_id, total_tasks, completed_tasks,
+                 workflow_spec_id, total_tasks, completed_tasks,
                  last_updated, is_review, title, study_id):
         self.id = id
         self.status = status
         self.next_task = next_task  # The next task that requires user input.
         self.navigation = navigation
         self.workflow_spec_id = workflow_spec_id
-        self.is_latest_spec = is_latest_spec
         self.total_tasks = total_tasks
         self.completed_tasks = completed_tasks
         self.last_updated = last_updated
@@ -231,7 +230,7 @@ class WorkflowApiSchema(ma.Schema):
     class Meta:
         model = WorkflowApi
         fields = ["id", "status", "next_task", "navigation",
-                  "workflow_spec_id", "is_latest_spec", "total_tasks", "completed_tasks",
+                  "workflow_spec_id", "total_tasks", "completed_tasks",
                   "last_updated", "is_review", "title", "study_id"]
         unknown = INCLUDE
 
@@ -242,7 +241,7 @@ class WorkflowApiSchema(ma.Schema):
     @marshmallow.post_load
     def make_workflow(self, data, **kwargs):
         keys = ['id', 'status', 'next_task', 'navigation',
-                'workflow_spec_id', 'is_latest_spec', "total_tasks", "completed_tasks",
+                'workflow_spec_id', "total_tasks", "completed_tasks",
                 "last_updated", "is_review", "title", "study_id"]
         filtered_fields = {key: data[key] for key in keys}
         filtered_fields['next_task'] = TaskSchema().make_task(data['next_task'])
