+++ crc/services/lookup_service.py
@@ -17,6 +17,7 @@ from crc.models.file import FileModel, FileDataModel, LookupFileModel, LookupDat
 from crc.models.ldap import LdapSchema
 from crc.models.workflow import WorkflowModel, WorkflowSpecDependencyFile
 from crc.services.file_service import FileService
+from crc.services.spec_file_service import SpecFileService
 from crc.services.ldap_service import LdapService
 from crc.services.workflow_processor import WorkflowProcessor
 
@@ -50,11 +51,12 @@ class LookupService(object):
         return LookupService.__get_lookup_model(workflow, spiff_task.task_spec.name, field.id)
 
     @staticmethod
-    def get_lookup_model_for_file_data(file_data: FileDataModel, value_column, label_column):
-        lookup_model = db.session.query(LookupFileModel).filter(LookupFileModel.file_data_model_id == file_data.id).first()
+    def get_lookup_model_for_file_data(file_id, file_name, value_column, label_column):
+        file_data = SpecFileService.get_reference_file_data(file_name)
+        lookup_model = db.session.query(LookupFileModel).filter(LookupFileModel.file_model_id == file_id).first()
         if not lookup_model:
             logging.warning("!!!! Making a very expensive call to update the lookup model.")
-            lookup_model = LookupService.build_lookup_table(file_data, value_column, label_column)
+            lookup_model = LookupService.build_lookup_table(file_id, file_name, file_data, value_column, label_column)
         return lookup_model
 
     @staticmethod
@@ -75,7 +77,6 @@ class LookupService(object):
                 is_current = db.session.query(WorkflowSpecDependencyFile). \
                     filter(WorkflowSpecDependencyFile.file_data_id == lookup_model.file_data_model_id).\
                     filter(WorkflowSpecDependencyFile.workflow_id == workflow.id).count()
-
 
         if not is_current:
             # Very very very expensive, but we don't know need this till we do.
@@ -131,7 +132,7 @@ class LookupService(object):
             file_name = field.get_property(Task.FIELD_PROP_SPREADSHEET_NAME)
             value_column = field.get_property(Task.FIELD_PROP_VALUE_COLUMN)
             label_column = field.get_property(Task.FIELD_PROP_LABEL_COLUMN)
-            latest_files = FileService().get_spec_data_files(workflow_spec_id=workflow_model.workflow_spec_id,
+            latest_files = SpecFileService().get_spec_data_files(workflow_spec_id=workflow_model.workflow_spec_id,
                                                            workflow_id=workflow_model.id,
                                                            name=file_name)
             if len(latest_files) < 1:
@@ -139,7 +140,10 @@ class LookupService(object):
             else:
                 data_dict = latest_files[0]
 
-            lookup_model = LookupService.build_lookup_table(data_dict, value_column, label_column,
+            file_id = data_dict['meta']['id']
+            file_name = data_dict['meta']['name']
+            file_data = data_dict['data']
+            lookup_model = LookupService.build_lookup_table(file_id, file_name, file_data, value_column, label_column,
                                                             workflow_model.workflow_spec_id, task_spec_id, field_id)
 
         #  Use the results of an LDAP request to populate enum field options
@@ -158,23 +162,23 @@ class LookupService(object):
         return lookup_model
 
     @staticmethod
-    def build_lookup_table(data_dict, value_column, label_column,
+    def build_lookup_table(file_id, file_name, file_data, value_column, label_column,
                            workflow_spec_id=None, task_spec_id=None, field_id=None):
         """ In some cases the lookup table can be very large.  This method will add all values to the database
          in a way that can be searched and returned via an api call - rather than sending the full set of
           options along with the form.  It will only open the file and process the options if something has
           changed.  """
-        if workflow_spec_id is not None:
-            file_data = data_dict['data']
-        else:
-            file_data = data_dict.data
+        # if workflow_spec_id is not None:
+        #     file_data = data_dict['data']
+        # else:
+        #     file_data = data_dict.data
         try:
             xlsx = ExcelFile(file_data, engine='openpyxl')
         # Pandas--or at least openpyxl, cannot read old xls files.
         # The error comes back as zipfile.BadZipFile because xlsx files are zipped xml files
         except BadZipFile:
             raise ApiError(code='excel_error',
-                           message=f"Error opening excel file {data_dict['meta']['name']}. You may have an older .xls spreadsheet. (file_model_id: {data_dict['meta']['file_model_id']} workflow_spec_id: {workflow_spec_id}, task_spec_id: {task_spec_id}, and field_id: {field_id})")
+                           message=f"Error opening excel file {file_name}. You may have an older .xls spreadsheet. (file_model_id: {file_id} workflow_spec_id: {workflow_spec_id}, task_spec_id: {task_spec_id}, and field_id: {field_id})")
         df = xlsx.parse(xlsx.sheet_names[0])  # Currently we only look at the fist sheet.
         df = df.convert_dtypes()
         df = df.loc[:, ~df.columns.str.contains('^Unnamed')] # Drop unnamed columns.
@@ -183,17 +187,17 @@ class LookupService(object):
 
         if value_column not in df:
             raise ApiError("invalid_enum",
-                           "The file %s does not contain a column named % s" % (data_dict['meta']['name'],
+                           "The file %s does not contain a column named % s" % (file_name,
                                                                                 value_column))
         if label_column not in df:
             raise ApiError("invalid_enum",
-                           "The file %s does not contain a column named % s" % (data_dict['meta']['name'],
+                           "The file %s does not contain a column named % s" % (file_name,
                                                                                 label_column))
 
         lookup_model = LookupFileModel(workflow_spec_id=workflow_spec_id,
                                        field_id=field_id,
                                        task_spec_id=task_spec_id,
-                                       file_data_model_id=data_dict['meta']['id'],
+                                       file_model_id=file_id,
                                        is_ldap=False)
 
         db.session.add(lookup_model)
